New authorization

+++ app/src/main/java/com/lukakordzaia/streamflow/ui/phone/home/ContinueWatchingInfoFragment.kt
@@ -8,8 +8,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.navArgs
+import com.lukakordzaia.streamflow.R
 import com.lukakordzaia.streamflow.databinding.DialogRemoveTitleBinding
 import com.lukakordzaia.streamflow.databinding.FragmentPhoneContinueWatchingInfoBinding
+import com.lukakordzaia.streamflow.network.LoadingState
 import com.lukakordzaia.streamflow.ui.baseclasses.BaseBottomSheet
 import com.lukakordzaia.streamflow.utils.AppConstants
 import com.lukakordzaia.streamflow.utils.EventObserver
@@ -48,10 +50,26 @@ class ContinueWatchingInfoFragment : BaseBottomSheet<FragmentPhoneContinueWatchi
             removeTitle.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
             removeTitle.setContentView(binding.root)
 
+            if (sharedPreferences.getLoginToken() != "") {
+                binding.title.text = resources.getString(R.string.remove_from_list_title)
+            }
+
            binding.continueButton.setOnClickListener {
-                homeViewModel.deleteContinueWatching(args.titleId)
-                removeTitle.dismiss()
-                dismiss()
+               if (sharedPreferences.getLoginToken() == "") {
+                   homeViewModel.deleteContinueWatching(args.titleId)
+                   removeTitle.dismiss()
+                   dismiss()
+               } else {
+                   homeViewModel.hideContinueWatchingLoader.observe(viewLifecycleOwner, {
+                       when (it.status) {
+                           LoadingState.Status.RUNNING -> {}
+                           LoadingState.Status.SUCCESS -> {
+                               removeTitle.dismiss()
+                               dismiss()
+                           }
+                       }
+                   })
+               }
             }
             binding.cancelButton.setOnClickListener {
                 removeTitle.dismiss()
